[Feat] : MemberRepository Reader와 Store로 구분 및 그에 따른 테스트코드 수정
