+++ src/main/java/org/programmers/VoucherManagement/member/infrastructure/MemberReaderRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-public interface MemberRepository {
+public interface MemberReaderRepository {
     /**
      * 전체 회원 조회
      *
@@ -28,27 +28,5 @@ public interface MemberRepository {
      * @param memberId
      * @return Optional<Member> - memberId값을 이용해 조회한 회원
      */
-    Optional<Member> findById(UUID memberId); //조회
-
-    /**
-     * db에 회원 저장
-     *
-     * @param member
-     * @return Member - 저장 완료한 회원
-     */
-    Member insert(Member member); //저장
-
-    /**
-     * db에 저장된 회원 수정
-     *
-     * @param member
-     */
-    void update(Member member); //수정
-
-    /**
-     * db에 저장된 회원 삭제
-     *
-     * @param memberId
-     */
-    void delete(UUID memberId); //삭제
+    Optional<Member> findById(UUID memberId);
 }
